+++ packages/@runlightyear/github/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @runlightyear/github
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @runlightyear/lightyear@0.6.0
+
 ## 0.3.0
 
 ### Minor Changes